simplify stringEnd
Instead of trying to reason about escaped/unescaped states, we
can unconditionally skip one character if we hit a backslash. If
the result of that would be skipping one past the end of a
string, nothing changes -- our index is too large either way and
we'll return -1. So we never need to check whether we're escaped,
we just handle each escape immediately upon seeing it. Note
that this works even if the escape sequence isn't actually a
backslash or a quote; all we care about, in stringEnd, is that
we don't treat a quote that was preceeded by an effective
backslash as the end of a string.

Trims a significant amount off the runtime for several of
the benchmarks, for me.